+++ src/rider/main/kotlin/com/github/rafaelldi/diagnosticsclientplugin/dialogs/CollectTracesDialog.kt
@@ -9,13 +9,15 @@ import com.intellij.ui.components.JBRadioButton
 import com.intellij.ui.components.fields.ExpandableTextField
 import com.intellij.ui.dsl.builder.*
 import com.jetbrains.rider.projectView.solutionDirectoryPath
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
 import javax.swing.JComponent
 
 class CollectTracesDialog(private val project: Project) : DialogWrapper(project) {
     private val model: CollectTracesModel =
         CollectTracesModel(
             project.solutionDirectoryPath.toString(),
-            "trace.nettrace",
+            getDefaultFilename(),
             StoppingType.AfterPeriod,
             30,
             TracingProfile.CpuSampling,
@@ -94,4 +96,11 @@ class CollectTracesDialog(private val project: Project) : DialogWrapper(project)
     fun getModel(): CollectTracesModel = model
 
     override fun getHelpId(): String = "com.github.rafaelldi.diagnosticsclientplugin.traces"
+
+    private fun getDefaultFilename(): String {
+        val current = LocalDateTime.now()
+        val formatter = DateTimeFormatter.ofPattern("yyyyMMdd_HHmmss")
+        val formatted = current.format(formatter)
+        return "trace_${formatted}.nettrace"
+    }
 }